Extract AutoComplete feature from RoomDetailFragment
